--- src/ModelRef.js
@@ -1,9 +1,0 @@
-// @flow
-
-export default class ModelRef {
-  name:string
-
-  constructor (name:string) {
-    this.name = name
-  }
-}
