--- src/mutation/index.js
@@ -1,10 +1,0 @@
-// @flow
-import addMutation from './methods/add'
-import deleteMutation from './methods/delete'
-import updateMutation from './methods/update'
-
-export default {
-  addMutation: addMutation,
-  deleteMutation: deleteMutation,
-  updateMutation: updateMutation
-}
