+++ src/schema/Schema.js
@@ -1,28 +1,24 @@
 // @flow
 
 import type { LinkedFieldConfig, QueryConfig, MutationConfig,
-  FieldType, ModelOptionConfig, HasOneConfig, BelongsToConfig, HasManyConfig,
-  BelongsToManyConfig, AssociationConfig} from './Definition'
+  FieldType, SchemaOptionConfig, HasOneConfig, BelongsToConfig, HasManyConfig,
+  BelongsToManyConfig, AssociationConfig} from '../Definition'
 
-/**
- * TODO
- * @example
- */
-export default class Model {
+export default class Schema {
   name:string
 
   config:{
     fields:{[id:string]: FieldType},
     links:{[id:string]:LinkedFieldConfig},
     associations:AssociationConfig,
-    options:ModelOptionConfig,
+    options:SchemaOptionConfig,
     queries:{[id:string]: QueryConfig},
     mutations:{[id:string]: MutationConfig},
     methods:{[id:string]: any},
     statics:{[id:string]: any}
   }
 
-  constructor (name:string, options:ModelOptionConfig = {}) {
+  constructor (name:string, options:SchemaOptionConfig = {}) {
     this.name = name
     this.config = {
       fields: {},
@@ -45,7 +41,7 @@ export default class Model {
    * Add the model base fields, and each field has a corresponding database column.
    * In default, each field generate a GraphQL field, unless it config with "hidden:true".
    */
-  fields (fields:{[id:string]: FieldType}):Model {
+  fields (fields:{[id:string]: FieldType}):Schema {
     this.config.fields = Object.assign(this.config.fields, fields)
     return this
   }
@@ -53,7 +49,7 @@ export default class Model {
   /**
    * Add the model link fields, and each link generate a GraphQL field but no corresponding database column.
    */
-  links (links:{[id:string]: LinkedFieldConfig}):Model {
+  links (links:{[id:string]: LinkedFieldConfig}):Schema {
     this.config.links = Object.assign(this.config.links, links)
     return this
   }
@@ -61,7 +57,8 @@ export default class Model {
   /**
    * Add the GraphQL query methods.
    */
-  queries (queries:{[string]:QueryConfig}):Model {
+  queries (queries:{[string]:QueryConfig}):Schema {
+    // TODO duplicate check
     this.config.queries = Object.assign(this.config.queries, queries)
     return this
   }
@@ -69,56 +66,61 @@ export default class Model {
   /**
    * Add the GraphQL mutataion methods.
    */
-  mutations (mutations:{[string]:MutationConfig}):Model {
+  mutations (mutations:{[string]:MutationConfig}):Schema {
+    // TODO duplicate check
     this.config.mutations = Object.assign(this.config.mutations, mutations)
     return this
   }
 
   /**
-   * Add instance method to current Model.
+   * Add instance method to current Schema.
    */
-  methods (methods:{[string]:any}):Model {
+  methods (methods:{[string]:any}):Schema {
     this.config.methods = Object.assign(this.config.methods, methods)
     return this
   }
 
   /**
-   * Add statics method to current Model.
+   * Add statics method to current Schema.
    */
-  statics (statics:{[string]:any}):Model {
+  statics (statics:{[string]:any}):Schema {
     this.config.statics = Object.assign(this.config.statics, statics)
     return this
   }
 
   /**
-   * Add {@link http://docs.sequelizejs.com/en/latest/docs/associations/#hasone|HasOne} relations to current Model.
+   * Add {@link http://docs.sequelizejs.com/en/latest/docs/associations/#hasone|HasOne} relations to current Schema.
    */
-  hasOne (config:HasOneConfig):Model {
+  hasOne (config:HasOneConfig):Schema {
     this.config.associations.hasOne.push(config)
     return this
   }
 
   /**
-   * Add {@link http://docs.sequelizejs.com/en/latest/docs/associations/#belongsto|BelongsTo} relations to current Model.
+   * Add {@link http://docs.sequelizejs.com/en/latest/docs/associations/#belongsto|BelongsTo} relations to current Schema.
    */
-  belongsTo (config:BelongsToConfig):Model {
+  belongsTo (config:BelongsToConfig):Schema {
     this.config.associations.belongsTo.push(config)
     return this
   }
 
   /**
-   * Add {@link http://docs.sequelizejs.com/en/latest/docs/associations/#one-to-many-associations|HasMany} relations to current Model.
+   * Add {@link http://docs.sequelizejs.com/en/latest/docs/associations/#one-to-many-associations|HasMany} relations to current Schema.
    */
-  hasMany (config:HasManyConfig):Model {
+  hasMany (config:HasManyConfig):Schema {
     this.config.associations.hasMany.push(config)
     return this
   }
 
   /**
-   * Add {@link http://docs.sequelizejs.com/en/latest/docs/associations/#belongs-to-many-associations|BelongsToMany} relations to current Model.
+   * Add {@link http://docs.sequelizejs.com/en/latest/docs/associations/#belongs-to-many-associations|BelongsToMany} relations to current Schema.
    */
-  belongsToMany (config:BelongsToManyConfig):Model {
+  belongsToMany (config:BelongsToManyConfig):Schema {
     this.config.associations.belongsToMany.push(config)
     return this
+  }
+
+  plugin<T> (plugin:(schema:Schema, options:T)=>void, options:T) {
+    plugin(this, options)
   }
 }
