+++ src/schema/index.js
@@ -1,0 +1,6 @@
+// @flow
+import Schema from './Schema'
+
+export default{
+  Schema: Schema
+}
